my_backend/requirements.txt
djangorestframework==3.9.4
djangorestframework-filters==0.11.1
djangorestframework-xml==1.4.0

